requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-pptx==0.6.23
openai==1.51.0
pydantic==2.9.2
python-dotenv==1.0.1
starlette==0.38.6